+++ packages/nimble-components/src/menu-button/tests/menu-button.spec.ts
@@ -1,12 +1,5 @@
 import { html } from '@microsoft/fast-element';
-import {
-    eventChange,
-    keyArrowDown,
-    keyArrowUp,
-    keyEnter,
-    keyEscape,
-    keySpace
-} from '@microsoft/fast-web-utilities';
+import { eventChange, keyEnter } from '@microsoft/fast-web-utilities';
 import { FoundationElement, Menu, MenuItem } from '@microsoft/fast-foundation';
 import { parameterizeSuite } from '@ni/jasmine-parameterized';
 import { fixture, Fixture } from '../../utilities/tests/fixture';
@@ -17,6 +10,7 @@ import {
     waitForUpdatesAsync
 } from '../../testing/async-helpers';
 import { createEventListener } from '../../utilities/tests/component';
+import { MenuButtonPageObject } from '../testing/menu-button.pageobject';
 
 class TestSlottedElement extends FoundationElement {}
 const composedTestSlottedElement = TestSlottedElement.compose({
@@ -97,12 +91,14 @@ describe('MenuButton', () => {
 
     describe('basic functionality', () => {
         let element: MenuButton;
+        let pageObject: MenuButtonPageObject;
         let connect: () => Promise<void>;
         let disconnect: () => Promise<void>;
 
         beforeEach(async () => {
             ({ element, connect, disconnect, parent } = await setup());
             createAndSlotMenu(element);
+            pageObject = new MenuButtonPageObject(element);
         });
 
         afterEach(async () => {
@@ -190,14 +186,14 @@ describe('MenuButton', () => {
         it('should not open menu when the toggle button is clicked if the element is disabled', async () => {
             element.disabled = true;
             await connect();
-            element.toggleButton!.control.click();
+            pageObject.clickMenuButton();
             expect(element.open).toBeFalse();
         });
 
         it('should close menu when toggle button is clicked while the menu is open', async () => {
             element.open = true;
             await connect();
-            element.toggleButton!.control.click();
+            pageObject.clickMenuButton();
             expect(element.open).toBeFalse();
         });
 
@@ -300,7 +296,7 @@ describe('MenuButton', () => {
                 oldState: false
             };
 
-            element.toggleButton!.control.click();
+            pageObject.clickMenuButton();
             await beforeToggleListener.promise;
             expect(beforeToggleListener.spy).toHaveBeenCalledTimes(1);
             const event = beforeToggleListener.spy.calls.first()
@@ -327,7 +323,7 @@ describe('MenuButton', () => {
                 oldState: true
             };
 
-            element.toggleButton!.control.click();
+            pageObject.clickMenuButton();
             await beforeToggleListener.promise;
             expect(beforeToggleListener.spy).toHaveBeenCalledTimes(1);
             const event = beforeToggleListener.spy.calls.first()
@@ -356,25 +352,17 @@ describe('MenuButton', () => {
     parameterizeSuite(menuSlotConfigurations, (suite, name, value) => {
         suite(`menu interaction with ${name}`, () => {
             let element: HTMLElement;
+            let pageObject: MenuButtonPageObject;
+            let menuButton: MenuButton;
             let connect: () => Promise<void>;
             let disconnect: () => Promise<void>;
-
-            async function openMenu(menuButton: MenuButton): Promise<void> {
-                if (menuButton.open) {
-                    return;
-                }
-
-                const toggleListener = createEventListener(
-                    menuButton,
-                    'toggle'
-                );
-                menuButton.open = true;
-                await toggleListener.promise;
-            }
 
             beforeEach(async () => {
                 ({ element, connect, disconnect, parent } = await value.setupFunction());
                 createAndSlotMenu(element);
+                await connect();
+                menuButton = value.getMenuButton(element);
+                pageObject = new MenuButtonPageObject(menuButton);
             });
 
             afterEach(async () => {
@@ -382,128 +370,90 @@ describe('MenuButton', () => {
             });
 
             it('should open the menu and focus first menu item when the toggle button is clicked', async () => {
-                await connect();
-                const menuButton = value.getMenuButton(element);
                 const toggleListener = createEventListener(
                     menuButton,
                     'toggle'
                 );
-                menuButton.toggleButton!.control.click();
+                pageObject.clickMenuButton();
                 expect(menuButton.open).toBeTrue();
                 await toggleListener.promise;
                 expect(document.activeElement).toEqual(menuItem1);
             });
 
             it("should open the menu and focus first menu item when 'Enter' is pressed while the toggle button is focused", async () => {
-                await connect();
-                const menuButton = value.getMenuButton(element);
                 const toggleListener = createEventListener(
                     menuButton,
                     'toggle'
                 );
-                const event = new KeyboardEvent('keypress', {
-                    key: keyEnter
-                } as KeyboardEventInit);
-                menuButton.toggleButton!.control.dispatchEvent(event);
+                pageObject.pressEnterKey();
                 expect(menuButton.open).toBeTrue();
                 await toggleListener.promise;
                 expect(document.activeElement).toEqual(menuItem1);
             });
 
             it("should open the menu and focus first menu item when 'Space' is pressed while the toggle button is focused", async () => {
-                await connect();
-                const menuButton = value.getMenuButton(element);
                 const toggleListener = createEventListener(
                     menuButton,
                     'toggle'
                 );
-                const event = new KeyboardEvent('keypress', {
-                    key: keySpace
-                } as KeyboardEventInit);
-                menuButton.toggleButton!.control.dispatchEvent(event);
+                pageObject.pressSpaceKey();
                 expect(menuButton.open).toBeTrue();
                 await toggleListener.promise;
                 expect(document.activeElement).toEqual(menuItem1);
             });
 
             it('should open the menu and focus first menu item when the down arrow is pressed while the toggle button is focused', async () => {
-                await connect();
-                const menuButton = value.getMenuButton(element);
                 const toggleListener = createEventListener(
                     menuButton,
                     'toggle'
                 );
-                const event = new KeyboardEvent('keydown', {
-                    key: keyArrowDown
-                } as KeyboardEventInit);
-                menuButton.toggleButton!.dispatchEvent(event);
+                pageObject.pressArrowDownKey();
                 expect(menuButton.open).toBeTrue();
                 await toggleListener.promise;
                 expect(document.activeElement).toEqual(menuItem1);
             });
 
             it('should open the menu and focus last menu item when the up arrow is pressed while the toggle button is focused', async () => {
-                await connect();
-                const menuButton = value.getMenuButton(element);
                 const toggleListener = createEventListener(
                     menuButton,
                     'toggle'
                 );
-                const event = new KeyboardEvent('keydown', {
-                    key: keyArrowUp
-                } as KeyboardEventInit);
-                menuButton.toggleButton!.dispatchEvent(event);
+                pageObject.pressArrowUpKey();
                 expect(menuButton.open).toBeTrue();
                 await toggleListener.promise;
                 expect(document.activeElement).toEqual(menuItem3);
             });
 
             it("should close the menu when pressing 'Escape'", async () => {
-                await connect();
-                const menuButton = value.getMenuButton(element);
-                await openMenu(menuButton);
+                await pageObject.openMenu();
 
-                const event = new KeyboardEvent('keydown', {
-                    key: keyEscape
-                } as KeyboardEventInit);
-                menuButton.region!.dispatchEvent(event);
+                pageObject.closeMenuWithEscape();
                 expect(menuButton.open).toBeFalse();
             });
 
             it("should focus the button when the menu is closed by pressing 'Escape'", async () => {
-                await connect();
-                const menuButton = value.getMenuButton(element);
-                await openMenu(menuButton);
+                await pageObject.openMenu();
 
-                const event = new KeyboardEvent('keydown', {
-                    key: keyEscape
-                } as KeyboardEventInit);
-                menuButton.region!.dispatchEvent(event);
+                pageObject.closeMenuWithEscape();
                 expect(document.activeElement).toEqual(element);
             });
 
             it('should close the menu when selecting a menu item by clicking it', async () => {
-                await connect();
-                const menuButton = value.getMenuButton(element);
-                await openMenu(menuButton);
+                await pageObject.openMenu();
 
                 menuItem1.click();
                 expect(menuButton.open).toBeFalse();
             });
 
             it('should focus the button when the menu is closed by selecting a menu item by clicking it', async () => {
-                await connect();
-                const menuButton = value.getMenuButton(element);
-                await openMenu(menuButton);
+                await pageObject.openMenu();
 
                 menuItem1.click();
                 expect(document.activeElement).toEqual(element);
             });
 
             it("should close the menu when selecting a menu item using 'Enter'", async () => {
-                await connect();
-                const menuButton = value.getMenuButton(element);
-                await openMenu(menuButton);
+                await pageObject.openMenu();
 
                 const event = new KeyboardEvent('keydown', {
                     key: keyEnter
@@ -513,9 +463,7 @@ describe('MenuButton', () => {
             });
 
             it("should focus the button when the menu is closed by selecting a menu item using 'Enter'", async () => {
-                await connect();
-                const menuButton = value.getMenuButton(element);
-                await openMenu(menuButton);
+                await pageObject.openMenu();
 
                 const event = new KeyboardEvent('keydown', {
                     key: keyEnter
@@ -530,10 +478,7 @@ describe('MenuButton', () => {
                     expect(document.activeElement).toEqual(element);
                     menuItemChangeEventHandled = true;
                 };
-
-                await connect();
-                const menuButton = value.getMenuButton(element);
-                await openMenu(menuButton);
+                await pageObject.openMenu();
 
                 menuItem1.addEventListener(eventChange, onMenuItemChange);
                 menuItem1.click();
@@ -542,9 +487,7 @@ describe('MenuButton', () => {
             });
 
             it('should not close the menu when clicking on a disabled menu item', async () => {
-                await connect();
-                const menuButton = value.getMenuButton(element);
-                await openMenu(menuButton);
+                await pageObject.openMenu();
 
                 menuItem1.disabled = true;
                 menuItem1.click();
@@ -554,8 +497,7 @@ describe('MenuButton', () => {
             it('should close the menu when the element loses focus', async () => {
                 const focusableElement = document.createElement('input');
                 parent.appendChild(focusableElement);
-                await connect();
-                const menuButton = value.getMenuButton(element);
+
                 // Start with the focus on the menu button so that it can lose focus later
                 menuButton.focus();
                 menuButton.open = true;
@@ -570,25 +512,17 @@ describe('MenuButton', () => {
     parameterizeSuite(menuSlotConfigurations, (suite, name, value) => {
         suite(`menu interaction without a ${name}`, () => {
             let element: HTMLElement;
+            let menuButton: MenuButton;
+            let pageObject: MenuButtonPageObject;
             let connect: () => Promise<void>;
             let disconnect: () => Promise<void>;
-
-            async function openMenu(menuButton: MenuButton): Promise<void> {
-                if (menuButton.open) {
-                    return;
-                }
-
-                const toggleListener = createEventListener(
-                    menuButton,
-                    'toggle'
-                );
-                menuButton.open = true;
-                await toggleListener.promise;
-            }
 
             beforeEach(async () => {
                 ({ element, connect, disconnect, parent } = await value.setupFunction());
                 // Unlike other tests, explicitly do not slot a menu in the parent element
+                await connect();
+                menuButton = value.getMenuButton(element);
+                pageObject = new MenuButtonPageObject(menuButton);
             });
 
             afterEach(async () => {
@@ -596,13 +530,11 @@ describe('MenuButton', () => {
             });
 
             it('should transition to the open state when the toggle button is clicked', async () => {
-                await connect();
-                const menuButton = value.getMenuButton(element);
                 const toggleListener = createEventListener(
                     menuButton,
                     'toggle'
                 );
-                menuButton.toggleButton!.control.click();
+                pageObject.clickMenuButton();
                 expect(menuButton.open).toBeTrue();
                 await toggleListener.promise;
                 expect(toggleListener.spy).toHaveBeenCalledTimes(1);
@@ -616,16 +548,11 @@ describe('MenuButton', () => {
             });
 
             it("should transition to the open state when 'Enter' is pressed while the toggle button is focused", async () => {
-                await connect();
-                const menuButton = value.getMenuButton(element);
                 const toggleListener = createEventListener(
                     menuButton,
                     'toggle'
                 );
-                const event = new KeyboardEvent('keypress', {
-                    key: keyEnter
-                } as KeyboardEventInit);
-                menuButton.toggleButton!.control.dispatchEvent(event);
+                pageObject.pressEnterKey();
                 expect(menuButton.open).toBeTrue();
                 await toggleListener.promise;
                 expect(toggleListener.spy).toHaveBeenCalledTimes(1);
@@ -639,16 +566,11 @@ describe('MenuButton', () => {
             });
 
             it("should transition to the open state when 'Space' is pressed while the toggle button is focused", async () => {
-                await connect();
-                const menuButton = value.getMenuButton(element);
                 const toggleListener = createEventListener(
                     menuButton,
                     'toggle'
                 );
-                const event = new KeyboardEvent('keypress', {
-                    key: keySpace
-                } as KeyboardEventInit);
-                menuButton.toggleButton!.control.dispatchEvent(event);
+                pageObject.pressSpaceKey();
                 expect(menuButton.open).toBeTrue();
                 await toggleListener.promise;
                 expect(toggleListener.spy).toHaveBeenCalledTimes(1);
@@ -662,16 +584,11 @@ describe('MenuButton', () => {
             });
 
             it('should transition to the open state when the down arrow is pressed while the toggle button is focused', async () => {
-                await connect();
-                const menuButton = value.getMenuButton(element);
                 const toggleListener = createEventListener(
                     menuButton,
                     'toggle'
                 );
-                const event = new KeyboardEvent('keydown', {
-                    key: keyArrowDown
-                } as KeyboardEventInit);
-                menuButton.toggleButton!.dispatchEvent(event);
+                pageObject.pressArrowDownKey();
                 expect(menuButton.open).toBeTrue();
                 await toggleListener.promise;
                 expect(toggleListener.spy).toHaveBeenCalledTimes(1);
@@ -685,16 +602,11 @@ describe('MenuButton', () => {
             });
 
             it('should transition to the open state when the up arrow is pressed while the toggle button is focused', async () => {
-                await connect();
-                const menuButton = value.getMenuButton(element);
                 const toggleListener = createEventListener(
                     menuButton,
                     'toggle'
                 );
-                const event = new KeyboardEvent('keydown', {
-                    key: keyArrowUp
-                } as KeyboardEventInit);
-                menuButton.toggleButton!.dispatchEvent(event);
+                pageObject.pressArrowUpKey();
                 expect(menuButton.open).toBeTrue();
                 await toggleListener.promise;
                 expect(toggleListener.spy).toHaveBeenCalledTimes(1);
@@ -708,26 +620,16 @@ describe('MenuButton', () => {
             });
 
             it("should transition to the closed state when pressing 'Escape'", async () => {
-                await connect();
-                const menuButton = value.getMenuButton(element);
-                await openMenu(menuButton);
+                await pageObject.openMenu();
 
-                const event = new KeyboardEvent('keydown', {
-                    key: keyEscape
-                } as KeyboardEventInit);
-                menuButton.region!.dispatchEvent(event);
+                pageObject.closeMenuWithEscape();
                 expect(menuButton.open).toBeFalse();
             });
 
             it("should focus the button when moving to the closed state by pressing 'Escape'", async () => {
-                await connect();
-                const menuButton = value.getMenuButton(element);
-                await openMenu(menuButton);
+                await pageObject.openMenu();
 
-                const event = new KeyboardEvent('keydown', {
-                    key: keyEscape
-                } as KeyboardEventInit);
-                menuButton.region!.dispatchEvent(event);
+                pageObject.closeMenuWithEscape();
                 expect(document.activeElement).toEqual(element);
             });
         });