+++ packages/spright-components/CHANGELOG.md
@@ -1,8 +1,89 @@
 # Change Log - @ni/spright-components
 
-This log was last generated on Thu, 13 Jun 2024 17:01:39 GMT and should not be manually modified.
+This log was last generated on Mon, 01 Jul 2024 20:14:06 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 1.0.30
+
+Mon, 01 Jul 2024 20:14:06 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v29.7.3
+
+## 1.0.29
+
+Thu, 27 Jun 2024 17:22:12 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v29.7.2
+
+## 1.0.28
+
+Wed, 26 Jun 2024 20:18:04 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v29.7.1
+
+## 1.0.27
+
+Tue, 25 Jun 2024 18:38:46 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v29.7.0
+
+## 1.0.26
+
+Thu, 20 Jun 2024 20:08:12 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v29.6.0
+
+## 1.0.25
+
+Tue, 18 Jun 2024 22:10:11 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v29.5.0
+
+## 1.0.24
+
+Tue, 18 Jun 2024 18:31:52 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v29.4.2
+
+## 1.0.23
+
+Mon, 17 Jun 2024 22:45:32 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v29.4.1
+
+## 1.0.22
+
+Mon, 17 Jun 2024 16:38:20 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v29.4.0
+
+## 1.0.21
+
+Mon, 17 Jun 2024 15:13:30 GMT
+
+### Patches
+
+- Remove unneeded rollup dependencies ([ni/nimble@4befcd5](https://github.com/ni/nimble/commit/4befcd5173fc83a2e3d352f5f69f1e45de4221e1))
+- Bump @ni/nimble-components to v29.3.7
 
 ## 1.0.20